New guest system
This commit implements a new guest system. Instead of using temporary invites that remove the user when they close discord, allow guests to stay in the server for a week. This avoids issues with guests repeatedly being removed after they vote on a PUG poll.

This commit also refactors the existing guest tagging code to the new module, and updates the guest link. Documentation will be updated when this commit is merged.